+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature: Add support for Clever auth providers.
 - Redesign: in Auth Modal display AuthProvider logos as selectable buttons when there are 2-4 AuthProviders instead of Combobox
+- Fix: Get loans to display on /loans from state.loans instead of state.collection.
 
 # 2.3.0